+++ CHANGELOG.md
@@ -14,6 +14,8 @@ The format is a modified version of [Keep a Changelog](https://keepachangelog.co
 
 ### Added
 
+- Anime library updates can populate filler marks from AnimeFillerList when enabled.
+
 ### Improved
 
 ### Fixed